Fix dependency updater
