+++ src/Maestro/Maestro.ContainerApp/Program.cs
@@ -6,6 +6,7 @@ using Maestro.ContainerApp;
 using Maestro.ContainerApp.Actors;
 using Maestro.ContainerApp.Queues;
 using Maestro.ContainerApp.Utils;
+using Maestro.Contracts;
 using Maestro.Data;
 using Microsoft.DotNet.DarcLib;
 using Microsoft.DotNet.GitHub.Authentication;
@@ -35,6 +36,7 @@ builder.Services.AddTransient<DarcRemoteMemoryCache>();
 builder.Services.AddTransient<Microsoft.DotNet.Services.Utility.ExponentialRetry>();
 builder.Services.AddTransient<IAzureDevOpsTokenProvider, AzureDevOpsTokenProvider>();
 builder.Services.AddTransient<IBarClient, MaestroBarClient>();
+builder.Services.AddTransient<IDependencyUpdater, DependencyUpdater>();
 builder.Services.AddTransient<IGitHubAppTokenProvider, GitHubAppTokenProvider>();
 builder.Services.AddTransient<IGitHubClientFactory, GitHubClientFactory>();
 builder.Services.AddTransient<IGitHubTokenProvider, GitHubTokenProvider>();
